feat(portal): adopt Nitro build with proper nx targets
Add @tanstack/nitro-v2-vite-plugin so the portal emits a self-contained,
self-listening Node server (dist/apps/developer-portal/server/index.mjs with
bundled deps) plus static assets. This replaces the hand-rolled server.mjs
adapter and the pnpm-deploy step in the Dockerfile, which existed only because
the app lacked Nitro.

Targets: build (vite build -> Nitro output, cache outputs wired), serve/dev
(vite), preview (vite preview), and a new start target (node server/index.mjs,
mirrors the container entrypoint). lint/typecheck/test remain inferred.

Resolve the dependency conflicts Nitro surfaced:
- h3 override was '>=2.0.1' (unresolvable; wrong major). Nitro v2 needs h3 1.x,
  so floor at '^1.15.11' (clears GHSA-4hxc-9384-m385).
- The start-server-core '>=1.167.30' security floor put two router-cores beside
  react-start 1.166, crashing SSR dehydrate. The old adapter hid this by never
  using start-server-core. Align the suite up to react-start 1.168.26 /
  react-router 1.170.16 (both above the CVE + malware-advisory patches).

Migrate the app for Start 1.168 breaking changes:
- Split server-only imports (@tanstack/react-start/server) into .server.ts
  modules; the .ts files keep the createServerFn wrappers + client-safe helpers.
- Rename createServerFn().inputValidator() -> .validator().
- Update test/mocks for the .server split and the react-router 1.170 internals.

Delete server.mjs + docker-entrypoint.sh; simplify the Dockerfile to copy the
Nitro output and run node server/index.mjs; switch the compose/Docker health
check to a TCP liveness probe (no more /healthz route). Gitignore + lint-ignore
Nitro's .nitro/.output artifacts.

Verified: build, typecheck (20 projects), lint, and 147 tests all pass; the
production server renders /login with SSR (200).
