+++ apps/developer-portal/Dockerfile
@@ -1,18 +1,11 @@
 # Developer Portal Dockerfile — production image.
 #
-# The portal is a TanStack Start app. Its Vite build emits a
-# framework-agnostic fetch handler (`server/server.js`, default export
-# `{ fetch }`) plus static client assets (`client/`) — it does NOT ship a
-# self-listening server. We therefore:
-#   1. build the app with Nx (`vite build`), and
-#   2. run it behind a tiny Node adapter (`server.mjs`) that pipes
-#      `node:http` requests into the fetch handler and serves the client
-#      assets.
-#
-# Dependency strategy mirrors the auth-server prod image: `pnpm deploy`
-# produces a self-contained tree whose node_modules contains real copies of
-# the workspace + external prod deps, so the build output's bare imports
-# (react, @tanstack/*) resolve at runtime without a custom resolver.
+# The portal is a TanStack Start app built with the Nitro v2 Vite plugin.
+# `nx build` emits a self-contained, self-listening Node server at
+# `dist/apps/developer-portal/server/index.mjs` (Nitro bundles its runtime
+# deps into `server/node_modules`) plus static assets under `public/`. So the
+# runner just copies that output and runs `node server/index.mjs` — no custom
+# adapter and no separate `pnpm deploy` step are needed.
 #
 # Requires BuildKit (default on Docker 23+) for the `--mount=type=cache`
 # pnpm-store mount. A sibling `Dockerfile.dockerignore` lifts the root
@@ -20,7 +13,7 @@
 #
 # syntax=docker/dockerfile:1.7
 
-# Stage 1: Workspace install + Nx build + pnpm deploy.
+# Stage 1: Workspace install + Nx build.
 FROM node:24-alpine AS builder
 WORKDIR /app
 RUN corepack enable
@@ -34,14 +27,9 @@ COPY apps ./apps
 RUN --mount=type=cache,id=pnpm,target=/pnpm/store \
     pnpm install --frozen-lockfile --store-dir /pnpm/store
 
-# Build the TanStack Start app -> dist/apps/developer-portal/{server,client}.
+# Build the TanStack Start + Nitro app. Output (server/ with bundled
+# node_modules, public/) lands in dist/apps/developer-portal.
 RUN pnpm nx build developer-portal
-
-# Produce a self-contained deploy bundle at /deploy containing the portal's
-# prod node_modules (real copies, not symlinks across the store).
-RUN --mount=type=cache,id=pnpm,target=/pnpm/store \
-    pnpm --filter @qauth-labs/developer-portal deploy --prod --legacy \
-    --store-dir /pnpm/store /deploy
 
 # Stage 2: Runner.
 FROM node:24-alpine AS runner
@@ -52,29 +40,20 @@ ENV NODE_ENV=production
 RUN addgroup -g 1001 -S nodejs && \
     adduser -S nodejs -u 1001
 
-# Self-contained prod node_modules from the deploy bundle.
-COPY --from=builder --chown=nodejs:nodejs /deploy/node_modules ./node_modules
-
-# Built app output: SSR fetch handler + static client assets.
+# Self-contained Nitro output: the Node server (with its bundled deps) plus
+# the static assets it serves. Nothing else is required at runtime.
 COPY --from=builder --chown=nodejs:nodejs /app/dist/apps/developer-portal/server ./server
-COPY --from=builder --chown=nodejs:nodejs /app/dist/apps/developer-portal/client ./client
-
-# Node adapter (container entrypoint) + entrypoint script.
-COPY --chown=nodejs:nodejs apps/developer-portal/server.mjs ./server.mjs
-COPY --chown=nodejs:nodejs apps/developer-portal/docker-entrypoint.sh ./docker-entrypoint.sh
-# BusyBox `sed -i` rewrites + re-owns to root; re-chown the single file so
-# USER nodejs can exec it. (Per-file chown, not -R — cheap.)
-RUN sed -i 's/\r$//' ./docker-entrypoint.sh && \
-    chmod +x ./docker-entrypoint.sh && \
-    chown nodejs:nodejs ./docker-entrypoint.sh
+COPY --from=builder --chown=nodejs:nodejs /app/dist/apps/developer-portal/public ./public
 
 USER nodejs
 
+# Nitro's Node server listens on PORT (default 3000); pin it for the image.
 ENV PORT=3001
-ENV HOST=0.0.0.0
 EXPOSE 3001
 
+# Liveness only: confirm the process is up and accepting connections, without
+# depending on SSR rendering or the upstream auth-server being reachable.
 HEALTHCHECK --interval=30s --timeout=10s --start-period=20s --retries=3 \
-  CMD node -e "require('http').get('http://localhost:3001/healthz', (r) => {process.exit(r.statusCode === 200 ? 0 : 1)}).on('error', () => process.exit(1))"
+  CMD node -e "require('net').connect(3001, 'localhost').on('connect', function () { process.exit(0); }).on('error', function () { process.exit(1); })"
 
-ENTRYPOINT ["/bin/sh", "/app/docker-entrypoint.sh"]
+CMD ["node", "server/index.mjs"]
