+++ apps/developer-portal/Dockerfile.dockerignore
@@ -15,6 +15,8 @@ node_modules
 dist
 .nx
 coverage
+.nitro
+.output
 
 # Testing
 *.test.ts
