--- apps/developer-portal/docker-entrypoint.sh
@@ -1,5 +1,0 @@
-#!/bin/sh
-set -e
-
-echo "Starting developer portal..."
-exec node server.mjs
