+++ apps/developer-portal/package.json
@@ -6,6 +6,7 @@
   "dependencies": {
     "@qauth-labs/ui": "workspace:*",
     "@tailwindcss/vite": "catalog:",
+    "@tanstack/nitro-v2-vite-plugin": "catalog:",
     "@tanstack/react-start": "catalog:",
     "@tanstack/react-router": "catalog:",
     "react": "catalog:",
