+++ apps/developer-portal/src/routes/_authed/dashboard.test.tsx
@@ -22,20 +22,20 @@ const PageComponent = Route.options.component as NonNullable<typeof Route.option
 const testUser = { sub: 'u1', email: 'dev@example.com', email_verified: true };
 
 // Minimal fake router for SSR-mode rendering via RouterContextProvider.
-// Route.useRouteContext() → useMatch({ from: '/_authed/dashboard' }) → useRouterState() →
-// reads router.state.matches when router.isServer === true.
+// Route.useRouteContext() → useMatch({ from: '/_authed/dashboard' }) →
+// router.stores.getRouteMatchStore(from).get() when router.isServer === true (1.170+).
+const dashboardMatch = {
+  id: '/_authed/dashboard',
+  routeId: '/_authed/dashboard',
+  search: {},
+  context: { user: testUser },
+};
 const fakeRouter = {
   isServer: true,
   options: {},
-  state: {
-    matches: [
-      {
-        id: '/_authed/dashboard',
-        routeId: '/_authed/dashboard',
-        search: {},
-        context: { user: testUser },
-      },
-    ],
+  stores: {
+    getRouteMatchStore: () => ({ get: () => dashboardMatch }),
+    matchStores: new Map(),
   },
 } as unknown as Parameters<typeof RouterContextProvider>[0]['router'];
