+++ apps/developer-portal/src/server/actions/api-keys.ts
@@ -1,16 +1,7 @@
 import { createServerFn } from '@tanstack/react-start';
-import { getRequestHeader, setResponseHeader } from '@tanstack/react-start/server';
 
-import {
-  type ApiKey,
-  type ApiKeyListData,
-  type ApiKeyWithSecret,
-  authServerClient,
-  type CreateApiKeyInput,
-  type Result,
-} from '../auth-server-client';
-import { readSessionCookie } from '../session-cookie';
-import { UNAUTHENTICATED } from './clients';
+import { type CreateApiKeyInput } from '../auth-server-client';
+import { createApiKeyHandler, listApiKeysHandler, revokeApiKeyHandler } from './api-keys.server';
 
 /**
  * Server functions for the static developer API-key surface (ADR-008 §6,
@@ -20,17 +11,6 @@ import { UNAUTHENTICATED } from './clients';
  * call is proxied with `Authorization: Bearer <token>`. The one-time plaintext
  * key (create) is returned to the browser once and is never persisted here.
  */
-
-/**
- * Read the developer's access token from the signed, HttpOnly session cookie.
- * Identical to the helper in `actions/clients.ts`; kept local so this module is
- * self-contained.
- */
-function readAccessToken(): string | null {
-  const session = readSessionCookie(getRequestHeader('cookie'));
-  if (!session || Date.now() >= session.expiresAt) return null;
-  return session.accessToken;
-}
 
 /** Defense-in-depth cap mirroring the auth-server's `name` limit (1..255). */
 const NAME_MAX = 255;
@@ -52,21 +32,8 @@ function normalizeClientId(data: unknown): { clientId: string } {
 // List
 // ---------------------------------------------------------------------------
 
-export async function listApiKeysHandler({
-  data,
-}: {
-  data: { clientId: string };
-}): Promise<Result<ApiKeyListData>> {
-  const token = readAccessToken();
-  if (!token) return UNAUTHENTICATED;
-  // Masked, per-developer data — keep it out of any shared/proxy cache for
-  // consistency with the rest of the client-management surface.
-  setResponseHeader('Cache-Control', 'no-store');
-  return authServerClient.listApiKeys(token, data.clientId);
-}
-
 export const listApiKeysFn = createServerFn({ method: 'GET' })
-  .inputValidator(normalizeClientId)
+  .validator(normalizeClientId)
   .handler(listApiKeysHandler);
 
 // ---------------------------------------------------------------------------
@@ -95,21 +62,8 @@ export function normalizeCreateApiKeyInput(data: unknown): {
   return { clientId: d['clientId'] as string, input: { name } };
 }
 
-export async function createApiKeyHandler({
-  data,
-}: {
-  data: { clientId: string; input: CreateApiKeyInput };
-}): Promise<Result<ApiKeyWithSecret>> {
-  const token = readAccessToken();
-  if (!token) return UNAUTHENTICATED;
-  const result = await authServerClient.createApiKey(token, data.clientId, data.input);
-  // The plaintext key is carried once — the browser must never cache it.
-  if (result.ok) setResponseHeader('Cache-Control', 'no-store');
-  return result;
-}
-
 export const createApiKeyFn = createServerFn({ method: 'POST' })
-  .inputValidator(normalizeCreateApiKeyInput)
+  .validator(normalizeCreateApiKeyInput)
   .handler(createApiKeyHandler);
 
 // ---------------------------------------------------------------------------
@@ -129,16 +83,6 @@ export function normalizeRevokeInput(data: unknown): { clientId: string; keyId: 
   return { clientId: d.clientId, keyId: d.keyId };
 }
 
-export async function revokeApiKeyHandler({
-  data,
-}: {
-  data: { clientId: string; keyId: string };
-}): Promise<Result<ApiKey>> {
-  const token = readAccessToken();
-  if (!token) return UNAUTHENTICATED;
-  return authServerClient.revokeApiKey(token, data.clientId, data.keyId);
-}
-
 export const revokeApiKeyFn = createServerFn({ method: 'POST' })
-  .inputValidator(normalizeRevokeInput)
+  .validator(normalizeRevokeInput)
   .handler(revokeApiKeyHandler);
